Rework redirectUri handling to be more robus
Instead of adding and removing individual redirectUris from the client,
which at times caused some race-condition issues, we now read all
CommonClientRedirectUri specs with the same `.spec.clientId` field and
combine their `.spec.redirectUri` fields into a list of unique URIs. The
Keycloak client is then updated with this list.

Note that this makes the operator authoritative when it comes to the
redirect uris of the Keycloak client. Any manual override done via the
admin GUI will be overriden on the next reconcile.